* master: (215 commits)
Make Python scripts in examples executable.
Added comment on new command.
Turn on error in not converged for 2d/gravity and 2d/subduction.
Small fix to script rewrite in packager.
Flip path replacement in packager.
Update copyright in pylith*.in files.
Adjust rewriting of scripts for Jenkins (different absolute path).
Added setup_cwd.sh for testing binary builds.
Updated mode (+x).
Updated PETSc version info.
Updated to PETSc 3.7.2.
Updated spatialdata version requirement.
Updated version number and user manual authors in preface.
Updated mode (+x).
Updated copyright (after merge).
Fixed typo in examples/2d/gravity/README.
Slight changes to README.
Updated mode (+x).
Updated mode.
Updated copyright to 2016.
...

Conflicts:
	libsrc/pylith/feassemble/ElasticityImplicitLgDeform.cc
	libsrc/pylith/materials/ElasticMaterial.cc

+++ libsrc/pylith/faults/TopologyOps.cc
@@ -9,7 +9,7 @@
 // This code was developed as part of the Computational Infrastructure
 // for Geodynamics (http://geodynamics.org).
 //
-// Copyright (c) 2010-2015 University of California, Davis
+// Copyright (c) 2010-2016 University of California, Davis
 //
 // See COPYING for license information.
 //
@@ -124,7 +124,7 @@ pylith::faults::TopologyOps::classifyCellsDM(PetscDM dmMesh,
       std::cout << "vReplaceCells.size: " << vReplaceCells.size() << std::endl;
       std::cout << "vNoReplaceCells.size: " << vNoReplaceCells.size() << std::endl;
     }
-    assert(classifySize < vReplaceCells.size() + vNoReplaceCells.size());
+    assert(size_t(classifySize) < vReplaceCells.size() + vNoReplaceCells.size());
     classifySize = vReplaceCells.size() + vNoReplaceCells.size();
     if (classifySize > classifyTotal) {
       std::ostringstream msg;
