+++ libsrc/pylith/meshio/MeshIOCubit.cc
@@ -9,7 +9,7 @@
 // This code was developed as part of the Computational Infrastructure
 // for Geodynamics (http://geodynamics.org).
 //
-// Copyright (c) 2010-2015 University of California, Davis
+// Copyright (c) 2010-2016 University of California, Davis
 //
 // See COPYING for license information.
 //
@@ -363,7 +363,7 @@ pylith::meshio::MeshIOCubit::_orientCells(int_array* const cells,
   PYLITH_METHOD_BEGIN;
 
   assert(cells);
-  assert(cells->size() == numCells*numCorners);
+  assert(cells->size() == size_t(numCells*numCorners));
 
   if (2 == meshDim && 4 == numCorners) { // QUAD4
     ; // do nothing
